# api/openapi.yaml
openapi: 3.0.3
info:
  title: Finanzas SaaS API
  version: 0.1.0
  description: >
    API REST de la aplicación de finanzas personales/SaaS. La autorización
    real la aplica PostgreSQL Row Level Security (ver supabase/migrations);
    esta API es una capa fina que reenvía el JWT del usuario a Supabase.
    Estado de implementación por recurso, ver tabla en README.md — esta
    especificación documenta el contrato objetivo completo, no solo lo ya
    codificado.
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    IngresoInput:
      type: object
      required: [espacio_id, descripcion, importe_esperado, fecha_prevista]
      properties:
        espacio_id: { type: string, format: uuid }
        categoria_id: { type: string, format: uuid, nullable: true }
        cuenta_id: { type: string, format: uuid, nullable: true }
        descripcion: { type: string }
        importe_esperado: { type: number, minimum: 0 }
        importe_real: { type: number, minimum: 0, nullable: true }
        moneda: { type: string, default: EUR }
        fecha_prevista: { type: string, format: date }
        fecha_cobrada: { type: string, format: date, nullable: true }
        periodicidad:
          type: string
          enum: [unico, semanal, mensual, bimestral, trimestral, semestral, anual]
        observaciones: { type: string, nullable: true }
...
security:
  - bearerAuth: []
paths:
  /ingresos:
    get:
      summary: Listar ingresos de un espacio financiero (implementado)
      parameters:
        - name: espacio_id
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: desde
          in: query
          schema: { type: string, format: date }
        - name: hasta
          in: query
          schema: { type: string, format: date }
        - name: categoria_id
          in: query
          schema: { type: string, format: uuid }
        - name: cuenta_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Lista de ingresos
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    post:
      summary: Crear un ingreso (implementado)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IngresoInput' }
      responses:
        '201':
          description: Ingreso creado
        '403':
          description: Sin permiso (bloqueado por RLS)
        '400':
          description: Validación fallida
  /ingresos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Obtener un ingreso (implementado)
      responses:
        '200': { description: OK }
        '404': { description: No encontrado o sin acceso }
    patch:
      summary: Actualizar un ingreso (implementado)
      responses:
        '200': { description: Actualizado }
        '403': { description: Sin permiso }
    delete:
      summary: Eliminar un ingreso (implementado)
      responses:
        '204': { description: Eliminado }
        '403': { description: Sin permiso }
...
  /activos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get: { summary: 'Obtener un activo (implementado)', responses: { '200': { description: OK } } }
    patch: { summary: 'Actualizar un activo (implementado)', responses: { '200': { description: OK } } }
    delete: { summary: 'Eliminar un activo (implementado)', responses: { '204': { description: Eliminado } } }
...